requirements.txt
aiohttp==3.6.2
discord.py==1.3.3
disputils==0.1.3
jikanpy==4.1.0
python-dotenv==0.13.0
